sale consignment improvement

+++ app/models/Sim_sale_consignments_model.php
@@ -77,6 +77,7 @@ class Sim_sale_consignments_model extends CI_Model
             for ($i = 0; $i < count($groupIds); $i++) {
 	            //Add sale consignment detail based on each group
 	            $simsInGroup = $this->getSimInGroup($groupIds[$i]);
+                $isInsertSuccess = true;
 		        if ($simsInGroup->num_rows() > 0) {
 		            foreach (($simsInGroup->result()) as $row) {
 		                $dataConsignDetail = array(
@@ -85,8 +86,16 @@ class Sim_sale_consignments_model extends CI_Model
 			            	'use_sale_consignment_id' => $insertedId,
 			            	'is_sale' => 0,
 	            		);
-	            	$this->db->insert("sale_consignment_detail", $dataConsignDetail);
+    	            	$satusInsert = $this->db->insert("sale_consignment_detail", $dataConsignDetail);
+                        $isInsertSuccess = $isInsertSuccess && $satusInsert;
 		            }
+
+                    //Update sim status to out stock after add sale consignment
+                    if($isInsertSuccess){
+                        foreach (($simsInGroup->result()) as $row) {
+                           $this->db->update("sim", array('is_in_stock' => 0), array('id' => $row->id));
+                        }
+                    }
 		        }
             }
             return true;
@@ -112,7 +121,10 @@ class Sim_sale_consignments_model extends CI_Model
         }
 
         if ($this->db->delete('sale_consignment_detail', array('use_group_id' => $gId))) {
-            return true;
+            //Turn sim back to stock if user delete group from sale consignment
+            if($this->db->update("sim", array('is_in_stock' => 1), array('use_sim_group_id' => $gId))){
+                 return true;
+            }
         }
         return FALSE;
     }
@@ -127,11 +139,26 @@ class Sim_sale_consignments_model extends CI_Model
             return FALSE;
         }
 
-        if ($this->db->delete('sale_consignment_detail', array('use_sale_consignment_id' => $id))) {
-            if($this->db->delete('sim_sale_consignments', array('id' => $id))){
-                return true;
+
+        $this->db->select('use_group_id')
+            ->where('use_sale_consignment_id',$id );
+        $q =  $this->db->get('sale_consignment_detail');
+        if ($q->num_rows() > 0) {
+            foreach (($q->result()) as $row) {
+                $data[] = $row;
+            }
+
+           foreach ($data as $g) {
+                $this->db->update("sim", array('is_in_stock' => 1), array('use_sim_group_id' => $g->use_group_id));
+           }
+
+            if ($this->db->delete('sale_consignment_detail', array('use_sale_consignment_id' => $id))) {
+                if($this->db->delete('sim_sale_consignments', array('id' => $id))){
+                    return true;
+                }
             }
         }
+
         return FALSE;
     }
